requirements/base.txt
scanpy==1.10.0
scvi-tools==1.1.0
anndata==0.10.0
numpy==1.26.0
pandas==2.1.0
scikit-learn==1.3.0
torch==2.1.0
matplotlib==3.8.0
seaborn==0.13.0
pyyaml==6.0
tqdm==4.66.0
pysradb==1.5.2
requests==2.31.0
